reorganize tests by splitting everything from auto-register into respective files
