+++ spec/modifiers/preprocess.spec.ts
@@ -54,4 +54,31 @@ describe('preprocess', () => {
       }
     );
   });
+
+  // TODO: This test should probably be made to work.
+  it.skip('can automatically register schemas in preprocess', () => {
+    const schema = z
+      .preprocess(arg => {
+        if (typeof arg === 'boolean') {
+          return arg;
+        }
+
+        if (typeof arg === 'string') {
+          if (arg === 'true') return true;
+          if (arg === 'false') return false;
+        }
+
+        return undefined;
+      }, z.boolean().openapi('PlainBoolean'))
+      .openapi('PreprocessedBoolean');
+
+    expectSchema([schema], {
+      PlainBoolean: {
+        type: 'boolean',
+      },
+      PreprocessedBoolean: {
+        type: 'boolean',
+      },
+    });
+  });
 });
