+++ spec/types/object.spec.ts
@@ -80,4 +80,103 @@ describe('object', () => {
       }
     );
   });
+
+  it('can automatically register object properties', () => {
+    const schema = z
+      .object({ key: z.string().openapi('Test') })
+      .openapi('Object');
+
+    expectSchema([schema], {
+      Test: {
+        type: 'string',
+      },
+
+      Object: {
+        type: 'object',
+        properties: {
+          key: {
+            $ref: '#/components/schemas/Test',
+          },
+        },
+        required: ['key'],
+      },
+    });
+  });
+
+  it('can automatically register extended parent properties', () => {
+    const schema = z.object({ id: z.number().openapi('NumberId') });
+
+    const extended = schema
+      .extend({
+        name: z.string().openapi('Name'),
+      })
+      .openapi('ExtendedObject');
+
+    expectSchema([extended], {
+      Name: {
+        type: 'string',
+      },
+
+      NumberId: {
+        type: 'number',
+      },
+
+      ExtendedObject: {
+        type: 'object',
+        properties: {
+          id: {
+            $ref: '#/components/schemas/NumberId',
+          },
+          name: {
+            $ref: '#/components/schemas/Name',
+          },
+        },
+        required: ['id', 'name'],
+      },
+    });
+  });
+
+  it('can automatically register extended schemas', () => {
+    const schema = z
+      .object({ id: z.string().openapi('StringId') })
+      .openapi('Object');
+
+    const extended = schema
+      .extend({
+        id: z.number().openapi('NumberId'),
+      })
+      .openapi('ExtendedObject');
+
+    expectSchema([extended], {
+      StringId: {
+        type: 'string',
+      },
+
+      NumberId: {
+        type: 'number',
+      },
+
+      Object: {
+        type: 'object',
+        properties: {
+          id: {
+            $ref: '#/components/schemas/StringId',
+          },
+        },
+        required: ['id'],
+      },
+
+      ExtendedObject: {
+        allOf: [
+          { $ref: '#/components/schemas/Object' },
+          {
+            type: 'object',
+            properties: {
+              id: { $ref: '#/components/schemas/NumberId' },
+            },
+          },
+        ],
+      },
+    });
+  });
 });
